Revert "Merge branch 'trunk' into task/add-sorting-to-document-merger"
This reverts commit f1be1fb2de26420c5f2843d3cab645407aa783b1, reversing
changes made to 98f9b96ff9244c8e738a1d120581e6e01b8e993a.

+++ src/Core/Framework/Test/Logging/LoggingServiceTest.php
@@ -5,14 +5,17 @@ namespace Shopware\Core\Framework\Test\Logging;
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Exception;
 use Monolog\Handler\TestHandler;
-use Monolog\Level;
 use Monolog\Logger;
 use PHPUnit\Framework\TestCase;
 use Shopware\Core\Content\MailTemplate\Service\Event\MailErrorEvent;
 use Shopware\Core\Content\Test\Flow\TestFlowBusinessEvent;
 use Shopware\Core\Framework\Context;
+use Shopware\Core\Framework\Event\BusinessEvent;
 use Shopware\Core\Framework\Event\FlowLogEvent;
+use Shopware\Core\Framework\Feature;
 use Shopware\Core\Framework\Log\LoggingService;
+use Shopware\Core\Framework\Test\Event\TestBusinessEvent;
+use Shopware\Core\Framework\Test\Logging\Event\LogAwareTestBusinessEvent;
 use Shopware\Core\Framework\Test\Logging\Event\LogAwareTestFlowEvent;
 use Shopware\Core\Framework\Test\TestCaseBase\IntegrationTestBehaviour;
 
@@ -36,17 +39,79 @@ class LoggingServiceTest extends TestCase
     }
 
     /**
-     * @throws Exception
+     * @throws \Doctrine\DBAL\Exception
      */
     protected function tearDown(): void
     {
+        if (!Feature::isActive('v6.5.0.0')) {
+            return;
+        }
         parent::tearDown();
         $connection = $this->getContainer()->get(Connection::class);
         $connection->executeStatement('DELETE FROM `log_entry`');
     }
 
+    public function testWriteBusinessEvents(): void
+    {
+        Feature::skipTestIfActive('v6.5.0.0', $this);
+
+        $handler = new TestHandler();
+        $logger = new Logger('testlogger', [$handler]);
+
+        $service = new LoggingService('test', $logger);
+
+        $service->logBusinessEvent(
+            new BusinessEvent(
+                TestBusinessEvent::EVENT_NAME,
+                new TestBusinessEvent($this->context)
+            )
+        );
+
+        $records = $handler->getRecords();
+
+        static::assertCount(1, $records);
+        $testRecord = $records[0];
+
+        static::assertEquals(TestBusinessEvent::EVENT_NAME, $testRecord['message']);
+        static::assertEquals('test', $testRecord['context']['environment']);
+        static::assertEquals(Logger::DEBUG, $testRecord['level']);
+        static::assertEmpty($testRecord['context']['additionalData']);
+    }
+
+    /**
+     * @depends testWriteBusinessEvents
+     */
+    public function testWriteLogAwareBusinessEvent(): void
+    {
+        Feature::skipTestIfActive('v6.5.0.0', $this);
+        $handler = new TestHandler();
+        $logger = new Logger('testlogger', [$handler]);
+
+        $service = new LoggingService(
+            'test',
+            $logger
+        );
+
+        $service->logBusinessEvent(
+            new BusinessEvent(
+                LogAwareTestBusinessEvent::EVENT_NAME,
+                new LogAwareTestBusinessEvent($this->context)
+            )
+        );
+
+        $records = $handler->getRecords();
+        static::assertCount(1, $records);
+        $testRecord = $records[0];
+
+        static::assertEquals(Logger::EMERGENCY, $testRecord['level']);
+        static::assertNotEmpty($testRecord['context']['additionalData']);
+        static::assertArrayHasKey('awesomekey', $testRecord['context']['additionalData']);
+        static::assertEquals($testRecord['context']['additionalData']['awesomekey'], 'awesomevalue');
+    }
+
     public function testWriteFlowEvents(): void
     {
+        Feature::skipTestIfInActive('v6.5.0.0', $this);
         $handler = new TestHandler();
         $logger = new Logger('testlogger', [$handler]);
 
@@ -61,14 +126,15 @@ class LoggingServiceTest extends TestCase
         static::assertCount(1, $records);
         $testRecord = $records[0];
 
-        static::assertEquals(TestFlowBusinessEvent::EVENT_NAME, $testRecord->message);
-        static::assertEquals('test', $testRecord->context['environment']);
-        static::assertEquals(Level::Debug, $testRecord->level);
-        static::assertEmpty($testRecord->context['additionalData']);
+        static::assertEquals(TestFlowBusinessEvent::EVENT_NAME, $testRecord['message']);
+        static::assertEquals('test', $testRecord['context']['environment']);
+        static::assertEquals(Logger::DEBUG, $testRecord['level']);
+        static::assertEmpty($testRecord['context']['additionalData']);
     }
 
     public function testWriteMailSendLogEvents(): void
     {
+        Feature::skipTestIfInActive('v6.5.0.0', $this);
         $handler = new TestHandler();
         $logger = new Logger('testlogger', [$handler]);
 
@@ -83,9 +149,9 @@ class LoggingServiceTest extends TestCase
         static::assertCount(1, $records);
         $testRecord = $records[0];
 
-        static::assertEquals(MailErrorEvent::NAME, $testRecord->message);
-        static::assertEquals('test', $testRecord->context['environment']);
-        static::assertEquals(Level::Error, $testRecord->level);
+        static::assertEquals(MailErrorEvent::NAME, $testRecord['message']);
+        static::assertEquals('test', $testRecord['context']['environment']);
+        static::assertEquals(Logger::ERROR, $testRecord['level']);
     }
 
     /**
@@ -93,6 +159,7 @@ class LoggingServiceTest extends TestCase
      */
     public function testWriteLogAwareFlowEvent(): void
     {
+        Feature::skipTestIfInActive('v6.5.0.0', $this);
         $handler = new TestHandler();
         $logger = new Logger('testlogger', [$handler]);
 
@@ -109,9 +176,9 @@ class LoggingServiceTest extends TestCase
         static::assertCount(1, $records);
         $testRecord = $records[0];
 
-        static::assertEquals(Level::Emergency, $testRecord->level);
-        static::assertNotEmpty($testRecord->context['additionalData']);
-        static::assertArrayHasKey('awesomekey', $testRecord->context['additionalData']);
-        static::assertEquals('awesomevalue', $testRecord->context['additionalData']['awesomekey']);
+        static::assertEquals(Logger::EMERGENCY, $testRecord['level']);
+        static::assertNotEmpty($testRecord['context']['additionalData']);
+        static::assertArrayHasKey('awesomekey', $testRecord['context']['additionalData']);
+        static::assertEquals($testRecord['context']['additionalData']['awesomekey'], 'awesomevalue');
     }
 }
